Moving main controller to separate `extension` package
