+++ pkg/operator/controller/extension/extension/controllerregistration/controllerregistration.go
