+++ pkg/operator/controller/extension/extension/reconciler_delete.go
